feat: allow admin to change max date setting
